Move entity options to different screen, and by default hide them
